19384 - The puppet agent fails to report resources as managed if they are capitalized
The resources.txt at first glance looks like it downcases the whole
resource but this would be an obvious bad idea and that is not what it
is doing.

then writing the managing_resource? method this was mistakenly assumed
to always downcase everything when instead it just downcases the type.

This commit adjusts the behaviour to match up with what is in
resources.txt