(maint) Don't require mcollective for all tasks
Because the test task is defined using an RSpec object, if rspec is installed,
many other libraries will also be loaded, which will cause failures on builders
trying to use the rake tasks. This commit lessens the risk by trying to load
mcollective, and rescuing the load error. If mcollective successfully loaded,
we continue to try to load the other libraries. This could still fail, but
failure is less likely. Really, the test task definition needs to be completely
reworked.