maint - make the spec tests a little more robust
Some tests were reliant on other tests loading/defining namespaces for them,
which meant you couldn't run the spec tests in isolation, only in the whole.

This commit adds missing require lines to ensure the aspects of the module
under test are loaded as expected.

Flushed out with this snippet:

    for spec in $(find spec -name \*_spec.rb); do echo $spec; bundle exec rspec $spec; done