maint - Next release will be 1.7.0
From here on I'm planning on augmenting our semantic versioning a
little, to steal a pattern I've seen in clojure projects.  This
steps for this are loosely:

1) when a new project is created it will have the written version
0.0.1.snapshot, this is what will be recorded in agent/PLUGIN_NAME.ddl.

2) when releasing the next version of PLUGIN_NAME a commit will be
made that changes the written version to 0.0.1.  This commit will
tagged with an annotated commit as 0.0.1.

3) immediately after tagging 0.0.1 the written version will be
incremented and have .snapshot re-appended, so you get 0.0.2.snapshot
(or 0.1.0.snapshot if semver conventions decree it)

Commits can then regularly continue, at the next release repeat
steps 2 and 3.

The prinicpal benefit of this is that it should be obvious when
looking at a checked out version of the project if you're looking
at the released state, or the advanced state of the project.