(MCOP-23) MCollective service plugin not working with puppet 3.5.x
A change introduced in Puppet 3.5.x has made it necessary to create a
basic Puppet context before we can lookup the Type object.

In the case where Puppet 3.5.x and the Puppet agent was present on a
system, Puppet would get configured in such a way that the service and
package agents would fail due to not being able to look up the relevant
Type.