MCOP-52 switch to Process.spawn on win32 systems
As reported in MCOP-52 and confirmed by much manual testing
Process.create doesn't seem to respect PATH, even though it should.

Here we use Process.spawn instead, which does behave correctly.  It
does force a requirement to use ruby 1.9.3 or later on windows, but
that's what is shipped in PE and in the Puppet FOSS MSI that the
unofficial MCollective-MSI builds on.