Enable config specification of signal_daemon
Previously, there was no mechanism by which users could choose what
method the runonce action would use to initiate a run. This caused
problems because the default on Posix platforms, when the daemon was
found to be running, was to signal the daemon. This method of triggering
a run disallowed any custom options to be passed, such as noop or
environment.

This commit exposes signal_daemon as a configuration parameter, but does
not change the default.