MCOP-309 cope with no response
The previous implementation of find_applying_nodes would assume
that a node had completed if they didn't respond to the puppet.status
action as the code was oriented around processing the results set
rather then the expected set of responses.

With this commit we change method to operate in terms of the hosts
we expected to hear from, and count the number of times we didn't
receive a response to the status action.  If we're unresponded to
we increment the :no_response, and on exceeding the threshold we
assume that host is not coming back.