MCOP-310 Document the default behavior of puppet.signal_dameon
Here we document the current default behavior of the signal_daemon
option, which matches the previous default unconfigurable value.

We tone down some of the polemic about the historic default, it's
not recommended given puppet 3.x's behaviour now allows a foreground
run alongside a daemon, but it's not going to eat any babies.