MCOP-330 extract making status hashes to 'make_status'
In 6e8b737f we added a new field to the running status hash used by
find_running_nodes, but overlooked that these hashes are also created
some distance away by the `runhosts` method.  This would result in
a comparison between nil and a number in certain codepaths.

With this commit we extract the creation of these status hashes into a
`make_status` method and make `runhosts` and `find_running_nodes` use it.