MCOP-557 disallow server option by default
This commit disallows the use of the `server` option by default as it
makes it possible for an attacker to execute arbitrary code on the node
by directing it to a compromised puppet master.
At the same time a new configuration flag - `allow_sever_override` - is
added to make it possible to explicitly allow the use of the option.