(MCOP-576) Consider AIO paths when determining agent command
AIO packaging does not add puppet to PATH so the default of `puppet
agent` does not work for modern users forcing every single user to edit
the config file just to use this.

This will check if AIO paths exist it will default to those otherwise
the old default of `puppet agent` is used.