(MCOP-580) Improve last_run_summary logs output
This improves the logs output from last_run_summary by adding details
about the logs such as time, source etc.

The old output was considered largely pointless and we decided to make
a breaking change to improve things.  The old 'parse_log' input is
removed thus forcing code that relied on the old behaviour to fail
with a DDL error rather than later fail in weird ways when the data format
has changed