(MCOP-585) Only use AIO location for platform, and if PATH fails
Previously default paths for the `puppet-agent` package were added to
allow this plugin to work out-of-the-box on those platforms. It did so
by hard-coding the paths and trying each of them in turn. This left a
hole for privilege escalation if not using `puppet-agent` packaging, and
on Windows in particular would allow anyone to create
`C:/opt/puppetlabs/bin/puppet` and co-opt the agent.

We aspire to prevent privilege escalation while retaining easy
out-of-the-box setup by
- Ensure default paths are only used on the platform where they apply
- Use PATH as the first option

On Windows, this is sufficient to prevent privilege escallation even if
not using the `puppet-agent` packaging, as `C:/Program Files` requires
Administrator access to modify. On other platforms, `/opt` is not a
privileged location, so we include a warning to explicitly set
`plugin.puppet.command` if not relying on PATH or using default
packaging.